Switch presses implemented
Switches are polling correctly and reading from Column to Row. Fixed pull down resistors on rows. Implemented debouncing maybe(?). TODO: map rest of the keys to the key map and layer switching button. Otherwise everything else should be good for the keyboard part of it unless we want to include menu control but I'm lazy lowkey.